requirements.txt
streamlit
pandas
pytrends
ollama
replicate
pytrends
python-dotenv
Pillow
requests
plotly
pandas